+++ src/fixtures/write-demo-output.js
@@ -1,0 +1,8 @@
+export const extractLoginResult = (output) => {
+  if (output.status === 'error') return 'error'
+
+  return output.output
+    ?.split('\n')
+    .filter((x) => !!x)
+    .pop()
+}
